test: add an e2e test to reproduce https://nvd.nist.gov/vuln/detail/CVE-2021-28235
